+++ apps/ymir-sdl3/src/app/app.hpp
@@ -115,7 +115,7 @@ private:
 
     void OpenLoadDiscDialog();
     void ProcessOpenDiscImageFileDialogSelection(const char *const *filelist, int filter);
-    bool LoadDiscImage(std::filesystem::path path);
+    bool LoadDiscImage(std::filesystem::path path, bool showErrorModal);
     void LoadRecentDiscs();
     void SaveRecentDiscs();