+++ libs/ymir-core/include/ymir/media/loader/loader_img_ccd_sub.hpp
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "loader_result.hpp"
+
 #include <ymir/media/disc.hpp>
 
 #include <filesystem>
@@ -50,6 +52,7 @@ namespace ymir::media::loader::ccd {
 // Returns true if loading all files succeeded.
 // If this function returns false, the Disc object is invalidated.
 // preloadToRAM specifies if the entire disc image should be preloaded into memory.
-bool Load(std::filesystem::path ccdPath, Disc &disc, bool preloadToRAM);
+// cbMsg is the callback for message reporting.
+bool Load(std::filesystem::path ccdPath, Disc &disc, bool preloadToRAM, CbLoaderMessage cbMsg);
 
 } // namespace ymir::media::loader::ccd