+++ libs/ymir-core/include/ymir/media/loader/loader_result.hpp
@@ -1,0 +1,17 @@
+#pragma once
+
+#include <functional>
+#include <string>
+
+namespace ymir::media {
+
+enum class MessageType {
+    InvalidFormat, // Parser has not detected valid magic signature
+    Error,         // Problem found parsing media; must be reported to the user
+    Debug,         // Debug messages, detailed parser logs
+};
+
+// Callback function for loader messages.
+using CbLoaderMessage = std::function<void(MessageType category, std::string message)>;
+
+} // namespace ymir::media